# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: LabelHub API
  version: 0.12.0
  description: LabelHub 本地演示 API，覆盖任务、模板、导入、草稿、提交、AI 审核、人工审核、终审和导出。
servers:
  - url: http://localhost:3000
    description: 本地开发 API
...
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      summary: 查询 Owner 任务
      parameters:
        - $ref: "#/components/parameters/OwnerId"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TaskStatus"
      responses:
        "200":
          description: 任务列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    post:
      tags: [Tasks]
      summary: 创建任务草稿
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTask"
      responses:
        "201":
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: 获取任务详情
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    patch:
      tags: [Tasks]
      summary: 更新任务基础信息
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTask"
      responses:
        "200":
          description: 更新后的任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /tasks/{id}/status:
    patch:
      tags: [Tasks]
      summary: 任务状态流转
      description: 支持草稿发布、暂停、恢复和结束，所有变更写入审计日志。
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/TaskStatus"
                actorId:
                  type: string
                reason:
                  type: string
                confirm:
                  type: boolean
      responses:
        "200":
          description: 更新后的任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /tasks/{id}/review-stage-config:
    patch:
      tags: [Tasks]
      summary: 配置审核阶段
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reviewStageConfig:
                  type: array
                  items:
                    type: string
                    enum: [RECHECK, FINAL]
                actorId:
                  type: string
      responses:
        "200":
          description: 更新后的任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /tasks/{id}/audit-logs:
    get:
      tags: [Tasks]
      summary: 查询任务审计日志
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 审计日志列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /templates:
    get:
      tags: [Templates]
      summary: 查询模板列表
      responses:
        "200":
          description: 模板列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    post:
      tags: [Templates]
      summary: 创建模板
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: 模板详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /templates/from-profile:
    post:
      tags: [Templates]
      summary: 从官方 DatasetProfile 创建模板
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                profile:
                  type: string
                  enum: [qa_quality, preference_compare, title_cleanup]
                actorId:
                  type: string
      responses:
        "201":
          description: 模板详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /templates/{id}:
    get:
      tags: [Templates]
      summary: 获取模板详情
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 模板详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    patch:
      tags: [Templates]
      summary: 更新模板 Schema
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "200":
          description: 模板详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /templates/{id}/publish:
    post:
      tags: [Templates]
      summary: 发布模板版本
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                versionName:
                  type: string
                actorId:
                  type: string
      responses:
        "201":
          description: 发布结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
...
  /exports:
    get:
      tags: [Exports]
      summary: 查询导出历史
      parameters:
        - name: taskId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 导出历史
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    post:
      tags: [Exports]
      summary: 创建导出任务
      parameters:
        - name: idempotency-key
          in: header
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [taskId, format, fieldMapping]
              properties:
                taskId:
                  type: string
                requestedById:
                  type: string
                format:
                  type: string
                  enum: [json, jsonl, csv, xlsx]
                includeReviews:
                  type: boolean
                fieldMapping:
                  type: array
                  items:
                    $ref: "#/components/schemas/FieldMapping"
                idempotencyKey:
                  type: string
      responses:
        "201":
          description: 导出任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /exports/{id}:
    get:
      tags: [Exports]
      summary: 查询导出详情
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 导出任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /exports/{id}/download:
    get:
      tags: [Exports]
      summary: 下载导出文件
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 文件下载
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /exports/{id}/retry:
    post:
      tags: [Exports]
      summary: 重试失败导出
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "201":
          description: 导出任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /tasks/{taskId}/export-preview:
    get:
      tags: [Exports]
      summary: 预览终审通过数据的导出字段映射
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - name: includeReviews
          in: query
          schema:
            type: boolean
        - name: fieldMapping
          in: query
          schema:
            type: string
            description: JSON 字符串形式的字段映射
      responses:
        "200":
          description: 导出预览，只包含 FINAL_APPROVED 数据
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
...
    OwnerId:
      name: ownerId
      in: query
      schema:
        type: string
  schemas:
    Envelope:
      type: object
      properties:
        data:
          description: 实际业务响应
        requestId:
          type: string
...
    DatasetKind:
      type: string
      enum: [qa_quality, preference_compare, generic_json]
    TaskStatus:
      type: string
      enum: [DRAFT, PUBLISHED, PAUSED, ENDED]
...
    CreateTask:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        richTextInstruction:
          type: string
        tags:
          type: array
          items:
            type: string
        rewardRule:
          type: string
        quota:
          type: integer
        deadline:
          type: string
          format: date-time
        distributionStrategy:
          type: string
          enum: [FIRST_COME_FIRST_SERVE, ASSIGNMENT, QUOTA_RACE]
        aiPreReviewEnabled:
          type: boolean
        aiRuleName:
          type: string
        templateId:
          type: string
        actorId:
          type: string
    TemplateInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        datasetKind:
          $ref: "#/components/schemas/DatasetKind"
        schema:
          $ref: "#/components/schemas/LabelHubSchema"
        actorId:
          type: string
    LabelHubSchema:
      type: object
      required: [schemaVersion, datasetKind, fields]
      properties:
        schemaVersion:
          type: string
        datasetKind:
          $ref: "#/components/schemas/DatasetKind"
        fields:
          type: array
          items:
            type: object
            additionalProperties: true
...
    FieldMapping:
      type: object
      properties:
        source:
          type: string
        target:
          type: string
        enabled:
          type: boolean